refactor: rename components to follow kebab-case
