+++ src/app/layout.tsx
@@ -2,8 +2,8 @@ import type { Metadata } from 'next';
 import { Geist, Geist_Mono } from 'next/font/google';
 import './globals.css';
 
-import Header from '@/components/header/Header';
-import Footer from '@/components/footer/Footer';
+import Header from '@/components/header/header';
+import Footer from '@/components/footer/footer';
 import { AuthProvider } from '@/context/AuthContext';
 const geistSans = Geist({
 	variable: '--font-geist-sans',
