+++ src/components/profile/profile-info.tsx
@@ -1,7 +1,7 @@
 import Image from 'next/image';
 import { mockedAvatar } from '@/assets';
 import { UserProfile } from '@/types';
-import ProfileFieldComponent from './ProfileField';
+import ProfileFieldComponent from './profile-field';
 import Button from '@/components/ui/Button/Button';
 import { PROFILE_FIELDS } from '@/constants/profile';
